generative/requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
requests==2.31.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
idna>=3.6
urllib3>=2.2.1